Named Pipe timeout and minor updates.
Added a threaded timeout to the write named pipe on application relaunch, in case the ReinstateIfRestarted method isn't called by the client application.
Added a few additional options to the WinForms sample application to aid debugging.
Tidied up FeedBuilder GUI and code (mainly post C# conversion bits and bobs).
Some refactoring and housekeeping.

Conflicts:
	NAppUpdate.sln
	src/NAppUpdate.Framework/Utils/FileDownloader.cs

+++ src/NAppUpdate.Updater/AppStart.cs
@@ -1,9 +1,9 @@
 ﻿using System;
-using System.IO;
 using System.Diagnostics;
+using System.IO;
 using System.Reflection;
-using System.Windows.Forms;
 using System.Threading;
+using System.Windows.Forms;
 using NAppUpdate.Framework;
 using NAppUpdate.Framework.Common;
 using NAppUpdate.Framework.Tasks;
@@ -26,51 +26,43 @@ namespace NAppUpdate.Updater
 
 			_logger = UpdateManager.Instance.Logger;
 			_args.ParseCommandLineArgs();
-			if (_args.ShowConsole)
-			{
+			if (_args.ShowConsole) {
 				_console = new ConsoleForm();
 				_console.Show();
 			}
 
 			Log("Starting to process cold updates...");
 
-			var workingDir = Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location);
-			if (_args.Log)
-			{
+			var workingDir = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
+			if (_args.Log) {
 				// Setup a temporary location for the log file, until we can get the DTO
 				logFile = Path.Combine(workingDir, @"NauUpdate.log");
 			}
 
-			try
-			{
+			try {
 				// Get the update process name, to be used to create a named pipe and to wait on the application
 				// to quit
 				string syncProcessName = _args.ProcessName;
-				if (string.IsNullOrEmpty(syncProcessName))
-					//Application.Exit();
-					throw new ArgumentException("The command line needs to specify the mutex of the program to update.", "args");
+				if (string.IsNullOrEmpty(syncProcessName)) //Application.Exit();
+					throw new ArgumentException("The command line needs to specify the mutex of the program to update.", "ar" + "gs");
 
 				Log("Update process name: '{0}'", syncProcessName);
 
 				// Load extra assemblies to the app domain, if present
 				var availableAssemblies = FileSystem.GetFiles(workingDir, "*.exe|*.dll", SearchOption.TopDirectoryOnly);
-				foreach (var assemblyPath in availableAssemblies)
-				{
+				foreach (var assemblyPath in availableAssemblies) {
 					Log("Loading {0}", assemblyPath);
 
-					if (assemblyPath.Equals(System.Reflection.Assembly.GetEntryAssembly().Location, StringComparison.InvariantCultureIgnoreCase)
-						|| assemblyPath.EndsWith("NAppUpdate.Framework.dll"))
-					{
+					if (assemblyPath.Equals(Assembly.GetEntryAssembly().Location, StringComparison.InvariantCultureIgnoreCase) || assemblyPath.EndsWith("NAppUpdate.Framework.dll")) {
 						Log("\tSkipping (part of current execution)");
 						continue;
 					}
 
-					try
-					{
-						var assembly = System.Reflection.Assembly.LoadFile(assemblyPath);
-					}
-					catch (System.BadImageFormatException ex)
-					{
+					try {
+// ReSharper disable UnusedVariable
+						var assembly = Assembly.LoadFile(assemblyPath);
+// ReSharper restore UnusedVariable
+					} catch (BadImageFormatException ex) {
 						Log("\tSkipping due to an error: {0}", ex.Message);
 					}
 				}
@@ -81,28 +73,22 @@ namespace NAppUpdate.Updater
 				// Make sure we start updating only once the application has completely terminated
 				Thread.Sleep(100); // hell, let's even wait a bit
 				bool createdNew;
-				using (var mutex = new Mutex(false, syncProcessName + "Mutex", out createdNew))
-				{
-					try
-					{
+				using (var mutex = new Mutex(false, syncProcessName + "Mutex", out createdNew)) {
+					try {
 						if (!createdNew) mutex.WaitOne();
-					}
-					catch (AbandonedMutexException)
-					{
+					} catch (AbandonedMutexException) {
 						// An abandoned mutex is exactly what we are expecting...
+					} finally {
 						Log("The application has terminated (as expected)");
 					}
 				}
 
 				bool updateSuccessful = true;
 
-				if (dto == null || dto.Configs == null)
-					throw new Exception("Invalid DTO received");
+				if (dto == null || dto.Configs == null) throw new Exception("Invalid DTO received");
 
 				if (dto.LogItems != null) // shouldn't really happen
-				{
 					_logger.LogItems.InsertRange(0, dto.LogItems);
-				}
 				dto.LogItems = _logger.LogItems;
 
 				// Get some required environment variables
@@ -112,22 +98,29 @@ namespace NAppUpdate.Updater
 				string backupFolder = dto.Configs.BackupFolder;
 				bool relaunchApp = dto.RelaunchApplication;
 
-				if (!string.IsNullOrEmpty(dto.AppPath))
-					logFile = Path.Combine(Path.GetDirectoryName(dto.AppPath), @"NauUpdate.log"); // now we can log to a more accessible location
+				if (!string.IsNullOrEmpty(dto.AppPath)) logFile = Path.Combine(Path.GetDirectoryName(dto.AppPath), @"NauUpdate.log"); // now we can log to a more accessible location
 
-				if (dto.Tasks == null || dto.Tasks.Count == 0)
-					throw new Exception("Could not find the updates list (or it was empty).");
+				if (dto.Tasks == null || dto.Tasks.Count == 0) throw new Exception("Could not find the updates list (or it was empty).");
 
 				Log("Got {0} task objects", dto.Tasks.Count);
 
+//This can be handy if you're trying to debug the updater.exe!
+//#if (DEBUG)
+//{  
+//                if (_args.ShowConsole) {
+//                    _console.WriteLine();
+//                    _console.WriteLine("Pausing to attach debugger.  Press any key to continue.");
+//                    _console.ReadKey();
+//                }
+ 
+//}
+//#endif
+
 				// Perform the actual off-line update process
-				foreach (var t in dto.Tasks)
-				{
+				foreach (var t in dto.Tasks) {
 					Log("Task \"{0}\": {1}", t.Description, t.ExecutionStatus);
 
-					if (t.ExecutionStatus != TaskExecutionStatus.RequiresAppRestart
-						&& t.ExecutionStatus != TaskExecutionStatus.RequiresPrivilegedAppRestart)
-					{
+					if (t.ExecutionStatus != TaskExecutionStatus.RequiresAppRestart && t.ExecutionStatus != TaskExecutionStatus.RequiresPrivilegedAppRestart) {
 						Log("\tSkipping");
 						continue;
 					}
@@ -135,76 +128,57 @@ namespace NAppUpdate.Updater
 					Log("\tExecuting...");
 
 					// TODO: Better handling on failure: logging, rollbacks
-					try
-					{
+					try {
 						t.ExecutionStatus = t.Execute(true);
-					}
-					catch (Exception ex)
-					{
+					} catch (Exception ex) {
 						Log(ex);
 						updateSuccessful = false;
 						t.ExecutionStatus = TaskExecutionStatus.Failed;
 					}
 
-					if (t.ExecutionStatus != TaskExecutionStatus.Successful)
-					{
-						Log("\tTask execution failed");
-						updateSuccessful = false;
-						break;
-					}
+					if (t.ExecutionStatus == TaskExecutionStatus.Successful) continue;
+					Log("\tTask execution failed");
+					updateSuccessful = false;
+					break;
 				}
 
-				if (updateSuccessful)
-				{
+				if (updateSuccessful) {
 					Log("Finished successfully");
 					Log("Removing backup folder");
-					if (Directory.Exists(backupFolder))
-						FileSystem.DeleteDirectory(backupFolder);
-				}
-				else
-				{
+					if (Directory.Exists(backupFolder)) FileSystem.DeleteDirectory(backupFolder);
+				} else {
 					MessageBox.Show("Update Failed");
 					Log(Logger.SeverityLevel.Error, "Update failed");
 				}
 
 				// Start the application only if requested to do so
-				if (relaunchApp)
-				{
+				if (relaunchApp) {
 					Log("Re-launching process {0} with working dir {1}", appPath, appDir);
 
-					var info = new ProcessStartInfo
-					           	{
-					           		UseShellExecute = true,
-					           		WorkingDirectory = appDir,
-					           		FileName = appPath,
-					           	};
-					
+					var info = new ProcessStartInfo {
+						UseShellExecute = true,
+						WorkingDirectory = appDir,
+						FileName = appPath,
+					};
+
 					var p = NauIpc.LaunchProcessAndSendDto(dto, info, syncProcessName);
-					if (p == null)
-						throw new UpdateProcessFailedException("Unable to relaunch application");
+					if (p == null) throw new UpdateProcessFailedException("Unable to relaunch application and/or send DTO");
 				}
 
 				Log("All done");
 				//Application.Exit();
-			}
-			catch (Exception ex)
-			{
+			} catch (Exception ex) {
 				// supressing catch because if at any point we get an error the update has failed
 				Log(ex);
-			}
-			finally
-			{
-				if (_args.Log)
-				{
+			} finally {
+				if (_args.Log) {
 					// at this stage we can't make any assumptions on correctness of the path
 					FileSystem.CreateDirectoryStructure(logFile, true);
 					_logger.Dump(logFile);
 				}
 
-				if (_args.ShowConsole)
-				{
-					if (_args.Log)
-					{
+				if (_args.ShowConsole) {
+					if (_args.Log) {
 						_console.WriteLine();
 						_console.WriteLine("Log file was saved to {0}", logFile);
 						_console.WriteLine();
@@ -222,22 +196,16 @@ namespace NAppUpdate.Updater
 		{
 			// Delete the updater EXE and the temp folder
 			Log("Removing updater and temp folder... {0}", tempFolder);
-			try
-			{
-				var info = new ProcessStartInfo
-				           	{
-				           		Arguments =
-				           			string.Format(@"/C ping 1.1.1.1 -n 1 -w 3000 > Nul & echo Y|del ""{0}\*.*"" & rmdir ""{0}""",
-				           			              tempFolder),
-				           		WindowStyle = ProcessWindowStyle.Hidden,
-				           		CreateNoWindow = true,
-				           		FileName = "cmd.exe"
-				           	};
+			try {
+				var info = new ProcessStartInfo {
+					Arguments = string.Format(@"/C ping 1.1.1.1 -n 1 -w 3000 > Nul & echo Y|del ""{0}\*.*"" & rmdir ""{0}""", tempFolder),
+					WindowStyle = ProcessWindowStyle.Hidden,
+					CreateNoWindow = true,
+					FileName = "cmd.exe"
+				};
 
 				Process.Start(info);
-			}
-			catch
-			{
+			} catch {
 				/* ignore exceptions thrown while trying to clean up */
 			}
 		}
@@ -252,8 +220,7 @@ namespace NAppUpdate.Updater
 			message = string.Format(message, args);
 
 			_logger.Log(severity, message);
-			if (_args.ShowConsole)
-				_console.WriteLine(message);
+			if (_args.ShowConsole) _console.WriteLine(message);
 
 			Application.DoEvents();
 		}
@@ -262,8 +229,7 @@ namespace NAppUpdate.Updater
 		{
 			_logger.Log(ex);
 
-			if (_args.ShowConsole)
-			{
+			if (_args.ShowConsole) {
 				_console.WriteLine("*********************************");
 				_console.WriteLine("   An error has occurred:");
 				_console.WriteLine("   " + ex);
